test(toh): migrate standalone app to functional route resolvers
