+++ packages/examples/tour-of-heroes-standalone/crisis-center/src/lib/crisis-center.routes.ts
@@ -2,7 +2,7 @@ import { Routes } from '@angular/router';
 import { canDeactivateGuard } from './can-deactivate.guard';
 import { CrisisCenterHomeComponent } from './crisis-center-home/crisis-center-home.component';
 import { CrisisCenterComponent } from './crisis-center/crisis-center.component';
-import { CrisisDetailResolverService } from './crisis-detail-resolver.service';
+import { crisisDetailResolver } from './crisis-detail.resolver';
 import { CrisisDetailComponent } from './crisis-detail/crisis-detail.component';
 import { CrisisListComponent } from './crisis-list/crisis-list.component';
 
@@ -20,7 +20,7 @@ export const crisisCenterRoutes: Routes = [
             component: CrisisDetailComponent,
             canDeactivate: [canDeactivateGuard],
             resolve: {
-              crisis: CrisisDetailResolverService,
+              crisis: crisisDetailResolver,
             },
           },
           {
